namespace-project/gloo-branch/pythonServer/requirements.txt
pika
pymongo
redis
opentelemetry-api==0.11b0
opentelemetry-sdk==0.11b0
opentelemetry-ext-jaeger
opentelemetry-ext-flask
opentelemetry-ext-requests
opentelemetry-ext-prometheus
prometheus_client